refactor: extract RateItems
