+++ packages/element3/src/components/Rate/src/rateItem.js
@@ -1,82 +1,11 @@
-import { ItemState } from './ItemState'
-import { reactive } from 'vue'
-
-const createRateItem = () => {
-  const item = reactive({
-    state: new ItemState(),
-    hover: false,
-    mouseOver() {
-      this.hover = true
-    },
-    mouseOut() {
-      this.hover = false
-    },
-    click() {
-      this.state.click()
-    },
-    reset() {
-      this.state.reset()
-    },
-    highlight() {
-      this.state.highlight()
-    },
-    putOut() {
-      this.state.putOut()
-    },
-    isStarOn() {
-      return this.state.isStarOn()
-    }
-  })
-
-  return item
-}
+import { RateItems } from './RateItems'
 
 export const useRateItem = (max) => {
-  const rateItems = reactive([])
+  const rateItems = new RateItems(max)
 
-  function init() {
-    for (let index = 0; index < max; index++) {
-      rateItems.push(createRateItem())
-    }
-  }
+  const click = (index) => rateItems.click(index)
+  const hover = (item, index) => rateItems.hover(item, index)
+  const putOut = (item) => rateItems.putOut(item)
 
-  function resetAllItems() {
-    rateItems.forEach((item) => {
-      item.reset()
-    })
-  }
-
-  function click(index) {
-    resetAllItems()
-
-    toStarOn(index, (item) => {
-      item.click()
-    })
-  }
-
-  function toStarOn(index, callback) {
-    rateItems.forEach((item, itemIndex) => {
-      if (itemIndex <= index) {
-        callback(item)
-      }
-    })
-  }
-
-  function hover(item, index) {
-    toStarOn(index, (item) => {
-      item.highlight()
-    })
-    item.mouseOver()
-  }
-
-  function putOut(item) {
-    rateItems.forEach((item) => {
-      item.putOut()
-    })
-    item.mouseOut()
-  }
-
-  init()
-
-  return { rateItems, click, hover, putOut }
+  return { rateItems: rateItems.getData(), click, hover, putOut }
 }
